refactor(events): Extract formatEventsToRender into event-helpers
